pyproject.toml
[tool.poetry]
name = "openai-compatible-api"
version = "0.1.0"
description = "OpenAI-compatible API for embeddings and reranking."
authors = ["Jules <jules@example.com>"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = {extras = ["standard"], version = ">=0.35.0,<0.36.0"}
pydantic = ">=2.11.9,<3.0.0"
pyyaml = ">=6.0.2,<7.0.0"
numpy = "^2.3.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
locust = "^2.40.4"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
